add modules and folder for tickets section
