+++ index.html
@@ -1349,7 +1349,7 @@
     <script src="/scripts/videoscripts/video.js" type="module"></script>
     <script src="/scripts/gallery.js" type="module"></script>
     <script src="/scripts/burgerMenu.js" type="module"></script>
-    <script src="/scripts/tickets.js" type="module"></script>
+    <script src="/scripts/ticketsscripts/tickets.js" type="module"></script>
 
 </body>
 </html>
